Entities DataTransformers refactoring/update
